+++ src/Entite/Admin_level.java
@@ -5,7 +5,7 @@
  */
 package Entite;
 
-import java.math.BigInteger;
+
 
 /**
  *
@@ -13,7 +13,7 @@ import java.math.BigInteger;
  */
 public class Admin_level {
     
-    protected BigInteger admin_level_id;
+    protected int admin_level_id;
     private int adm_cod ;
 
     public Admin_level(int adm_cod) {
@@ -28,11 +28,11 @@ public class Admin_level {
         this.adm_cod = adm_cod;
     }
 
-    public BigInteger getAdmin_level_id() {
+    public int getAdmin_level_id() {
         return admin_level_id;
     }
 
-    public void setAdmin_level_id(BigInteger admin_level_id) {
+    public void setAdmin_level_id(int admin_level_id) {
         this.admin_level_id = admin_level_id;
     }
